Consolidate BEQ UGens
