--- supriya/ugens/BEQSuite.py
@@ -1,11 +1,0 @@
-from supriya.ugens.Filter import Filter
-
-
-class BEQSuite(Filter):
-    """
-    Abstract base class of all BEQSuite UGens.
-    """
-
-    ### CLASS VARIABLES ###
-
-    __documentation_section__ = "Filter UGens"
